More testing infra work
