+++ x/superfluid/keeper/keeper_test.go
@@ -5,11 +5,16 @@ import (
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/simapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/osmosis-labs/osmosis/v7/app"
 	"github.com/osmosis-labs/osmosis/v7/x/superfluid/types"
 	"github.com/stretchr/testify/suite"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+
+	"github.com/osmosis-labs/osmosis/v7/x/gamm/pool-models/balancer"
+	gammtypes "github.com/osmosis-labs/osmosis/v7/x/gamm/types"
+	"github.com/tendermint/tendermint/crypto/ed25519"
 )
 
 type KeeperTestSuite struct {
@@ -34,6 +39,42 @@ func (suite *KeeperTestSuite) SetupDefaultPool() {
 	bondDenom := suite.app.StakingKeeper.BondDenom(suite.ctx)
 	poolId := suite.createGammPool([]string{bondDenom, "foo"})
 	suite.Require().Equal(poolId, uint64(1))
+}
+
+// CreateRandomAccounts is a function return a list of randomly generated AccAddresses
+func CreateRandomAccounts(numAccts int) []sdk.AccAddress {
+	testAddrs := make([]sdk.AccAddress, numAccts)
+	for i := 0; i < numAccts; i++ {
+		pk := ed25519.GenPrivKey().PubKey()
+		testAddrs[i] = sdk.AccAddress(pk.Address())
+	}
+
+	return testAddrs
+}
+
+func (suite *KeeperTestSuite) createGammPool(denoms []string) uint64 {
+	coins := suite.app.GAMMKeeper.GetParams(suite.ctx).PoolCreationFee
+	poolAssets := []gammtypes.PoolAsset{}
+	for _, denom := range denoms {
+		coins = coins.Add(sdk.NewInt64Coin(denom, 1000000000000000000))
+		poolAssets = append(poolAssets, gammtypes.PoolAsset{
+			Weight: sdk.NewInt(100),
+			Token:  sdk.NewCoin(denom, sdk.NewInt(1000000000000000000)),
+		})
+	}
+
+	acc1 := CreateRandomAccounts(1)[0]
+	err := simapp.FundAccount(suite.app.BankKeeper, suite.ctx, acc1, coins)
+	suite.Require().NoError(err)
+
+	poolId, err := suite.app.GAMMKeeper.CreateBalancerPool(
+		suite.ctx, acc1, balancer.PoolParams{
+			SwapFee: sdk.NewDecWithPrec(1, 2),
+			ExitFee: sdk.NewDecWithPrec(1, 2),
+		}, poolAssets, "")
+	suite.Require().NoError(err)
+
+	return poolId
 }
 
 func TestKeeperTestSuite(t *testing.T) {
